pyproject.toml
[project]
name = "data-intelligence-crawling"
version = "0.1.0"
description = "Web data collection with OpenSpec-driven development"
readme = "AGENTS.md"
requires-python = ">=3.11,<3.13"
authors = [{ name = "Jong" }]

dependencies = [
  "beautifulsoup4",
  "google-api-python-client",
  "google-auth",
  "google-auth-httplib2",
  "google-auth-oauthlib",
  "httpx",
  "lxml",
  "pandas",
  "pydantic",
  "python-dotenv",
  "rich",
  "tenacity",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]


[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[dependency-groups]
dev = [
    "pyyaml>=6.0.3",
]
